Reflection ok, vec2 removed, new map
I've fixed the artefact in the diffuse light.

The reflection now working as intended.

I've remove the vec2 struct because it's unused.

I've view a new bug, the y axis is reverse !

I've create a new map with 3 spheres, two sphere facing each other with the light beetween,
and the last one a really big one at the bottom to make a reference for movement.