Texturing arrived !
I've removed the fisheye corrector function because raytracing haven't fisheye
and it make the display wird.

I've re-add the vec2 because now it used in texturing.

I've add the compute of the pixel position on texture / normal map.
This position is between 0 and 1. With this we can handle any size of texture and normal map.

I've done the mapping functions for sphere.
Now sphere can have it base color like before, a checkerboard or a texture !
If the color_type is on texture and the texture doesn't exist, it will display
a checkerboard black and pink, like Minecraft.