Cubes join the battle !
I've done the parsing, the intersections, normal computing, normal map, texturing, reflection
and refraction.

We need to update the parser because I've return an error if the two axis given for cube isn't valid.
We need to modify the normal mapping beacause we don't have the correct base for the normal get.