Normal mapping for cube ok
I've done the normal mapping for cubes.

The normal mapping have a problem. When we change the normal according to the normal map,
if the face isn't facing a light, we can see some light effet, because the new normal
isn't in the same way of the original normal.

We need to free textures used too.