Update of parsing
Now we can specify if we want a color, a checkerboard, a texture with or without normal !
It work like this :

Take an object identifier obj
obj    -> the object like before
obj-ck -> the object with a generate checkerboard
obj-t  -> the object with a texture
obj-tn -> the object with a texture and a normal map

I've finish the cube parsing too.
Now when the cube's axis aren't valid, we have a message.

I've create a problem, the empty argument doesn't work any more.