Cone intersection normed
I've normed the cone intersection and optimazed it a bit.
Now cones have there ratio, which is the radius divide by height.
This ratio is squarred and add to 1 to make it usefull for intersection check.

I've normed 2/3 files two with some 'too many tabs' or 'mixed space and tab'